Some updares on the project
I did the search mothed. the program is able to click customers and see what cars are rented by them, also when the rent car button is clicked for the customer table and goes to the rent car table it displays Customers name at the table. and creating check boxes in the find car. Also you can add customer and cars.