Update for the project
The searching buttons are working. It is able to click customers and see what cars are rented by them. Also, when the rent car button is clicked for the customer table and goes to the rent car table it displays customers name at the table. There is check boxes in the find car.